+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 ### Bug Fixes
 
 - fix grouping routes by which bundle they're from
+- fix registration of resource method routes so the order is deterministic
 - fix `ConfigureAppHook` to load configs from every bundle in the hierarchy
 - fix passing explicit rule overrides to `routes.resource`
 - fix automatic endpoint names for resource routes